backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database dependencies
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.12.1

# Google APIs
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Facebook/Meta APIs
facebook-business==18.0.0

# WooCommerce API
woocommerce==3.0.0

# Data processing
pandas==2.1.3
numpy==1.25.2
python-dateutil==2.8.2

# Background tasks & scheduling
celery==5.3.4
redis==5.0.1
flower==2.0.1

# AI & ML
openai==1.3.7
scikit-learn==1.3.2

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Monitoring & logging
structlog==23.2.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
qrcode[pil]==7.4.2
pyotp==2.9.0 